+++ ledger/__init__.py
@@ -2,11 +2,12 @@
 
 from ledger.base import LedgerFilters, LedgerWriter, SearchResult
 from ledger.exceptions import LedgerError, LedgerReadError, LedgerWriteError
-from ledger.models import LedgerEntry, LedgerSource, LedgerStatus
+from ledger.models import LedgerEntry, LedgerSource, LedgerStatus, LedgerSummary
 from ledger.sqlite_writer import SQLiteLedgerWriter
 
 __all__ = [
     "LedgerEntry",
+    "LedgerSummary",
     "LedgerError",
     "LedgerFilters",
     "LedgerReadError",